+++ apps/meteor/app/api/server/middlewares/metrics.ts
@@ -2,22 +2,35 @@ import type { MiddlewareHandler } from 'hono';
 import type { Summary } from 'prom-client';
 
 import type { CachedSettings } from '../../../settings/server/CachedSettings';
-import type { APIClass } from '../api';
+import type { APIClass } from '../ApiClass';
 
 export const metricsMiddleware =
-	(api: APIClass, settings: CachedSettings, summary: Summary): MiddlewareHandler =>
+	({
+		basePathRegex,
+		api,
+		settings,
+		summary,
+	}: {
+		basePathRegex?: RegExp;
+		api: APIClass;
+		settings: CachedSettings;
+		summary: Summary;
+	}): MiddlewareHandler =>
 	async (c, next) => {
-		const { method, path } = c.req;
-
-		const rocketchatRestApiEnd = summary.startTimer({
-			method,
-			version: api.version,
-			...(settings.get('Prometheus_API_User_Agent') && { user_agent: c.req.header('user-agent') }),
-			entrypoint: path.startsWith('method.call') ? decodeURIComponent(c.req.url.slice(8)) : path,
-		});
+		const rocketchatRestApiEnd = summary.startTimer();
 
 		await next();
+
+		const { method, path, routePath } = c.req;
+
+		// get rid of the base path (i.e.: /api/v1/)
+		const entrypoint = basePathRegex ? routePath.replace(basePathRegex, '') : routePath;
+
 		rocketchatRestApiEnd({
 			status: c.res.status,
+			method: method.toLowerCase(),
+			version: api.version,
+			...(settings.get('Prometheus_API_User_Agent') && { user_agent: c.req.header('user-agent') }),
+			entrypoint: basePathRegex && entrypoint.startsWith('method.call') ? decodeURIComponent(path.replace(basePathRegex, '')) : entrypoint,
 		});
 	};