+++ apps/meteor/client/components/Contextualbar/ContextualbarScrollableContent.tsx
@@ -1,4 +1,5 @@
 import { Margins } from '@rocket.chat/fuselage';
+import { FeaturePreview, FeaturePreviewOff, FeaturePreviewOn } from '@rocket.chat/ui-client';
 import type { ComponentProps } from 'react';
 import { forwardRef, memo } from 'react';
 
@@ -7,9 +8,18 @@ import { PageScrollableContent } from '../Page';
 const ContextualbarScrollableContent = forwardRef<HTMLElement, ComponentProps<typeof PageScrollableContent>>(
 	function ContextualbarScrollableContent({ children, ...props }, ref) {
 		return (
-			<PageScrollableContent p={24} {...props} ref={ref}>
-				<Margins blockEnd={16}>{children}</Margins>
-			</PageScrollableContent>
+			<FeaturePreview feature='newNavigation'>
+				<FeaturePreviewOff>
+					<PageScrollableContent p={24} {...props} ref={ref}>
+						<Margins blockEnd={16}>{children}</Margins>
+					</PageScrollableContent>
+				</FeaturePreviewOff>
+				<FeaturePreviewOn>
+					<PageScrollableContent paddingInline={16} {...props} ref={ref}>
+						<Margins blockEnd={16}>{children}</Margins>
+					</PageScrollableContent>
+				</FeaturePreviewOn>
+			</FeaturePreview>
 		);
 	},
 );