+++ apps/meteor/client/omnichannel/tags/TagsPage.tsx
@@ -1,4 +1,5 @@
 import { Button, ButtonGroup } from '@rocket.chat/fuselage';
+import { useEffectEvent } from '@rocket.chat/fuselage-hooks';
 import { useRouter, useTranslation, useRouteParameter } from '@rocket.chat/ui-contexts';
 
 import TagEdit from './TagEdit';
@@ -13,6 +14,8 @@ const TagsPage = () => {
 	const context = useRouteParameter('context');
 	const id = useRouteParameter('id');
 
+	const handleCloseContextualbar = useEffectEvent(() => router.navigate('/omnichannel/tags'));
+
 	return (
 		<Page flexDirection='row'>
 			<Page>
@@ -26,9 +29,9 @@ const TagsPage = () => {
 				</PageContent>
 			</Page>
 			{context && (
-				<ContextualbarDialog>
-					{context === 'edit' && id && <TagEditWithData tagId={id} />}
-					{context === 'new' && <TagEdit />}
+				<ContextualbarDialog onClose={handleCloseContextualbar}>
+					{context === 'edit' && id && <TagEditWithData tagId={id} onClose={handleCloseContextualbar} />}
+					{context === 'new' && <TagEdit onClose={handleCloseContextualbar} />}
 				</ContextualbarDialog>
 			)}
 		</Page>