+++ ee/packages/omnichannel-services/jest.config.ts
@@ -3,4 +3,5 @@ import type { Config } from 'jest';
 
 export default {
 	preset: server.preset,
+	testMatch: ['<rootDir>/src/**/*.spec.(ts|js|mjs)'],
 } satisfies Config;