+++ packages/livechat/src/components/Screen/ScreenProvider.tsx
@@ -1,10 +1,9 @@
 import type { FunctionalComponent } from 'preact';
 import { createContext } from 'preact';
-import { useCallback, useContext, useEffect, useState } from 'preact/hooks';
+import { useContext, useEffect, useState } from 'preact/hooks';
 import { parse } from 'query-string';
 
 import { isActiveSession } from '../../helpers/isActiveSession';
-import { createOrUpdateGuest, evaluateChangesAndLoadConfigByFields } from '../../lib/hooks';
 import { loadConfig } from '../../lib/main';
 import { parentCall } from '../../lib/parentCall';
 import { loadMessages } from '../../lib/room';
@@ -41,6 +40,7 @@ export type ScreenContextValue = {
 		background?: string;
 		hideGuestAvatar?: boolean;
 		hideAgentAvatar?: boolean;
+		hideExpandChat?: boolean;
 	};
 };
 
@@ -51,6 +51,7 @@ export const ScreenContext = createContext<ScreenContextValue>({
 		iconColor: '',
 		hideAgentAvatar: false,
 		hideGuestAvatar: true,
+		hideExpandChat: false,
 	},
 	notificationsEnabled: true,
 	minimized: true,
@@ -63,21 +64,11 @@ export const ScreenContext = createContext<ScreenContextValue>({
 } as ScreenContextValue);
 
 export const ScreenProvider: FunctionalComponent = ({ children }) => {
-	const {
-		dispatch,
-		config,
-		sound,
-		minimized = true,
-		undocked,
-		expanded = false,
-		alerts,
-		modal,
-		iframe,
-		...store
-	} = useContext(StoreContext);
+	const store = useContext(StoreContext);
+	const { token, dispatch, config, sound, minimized = true, undocked, expanded = false, alerts, modal, iframe, customFieldsQueue } = store;
 	const { department, name, email } = iframe.guest || {};
-	const { color, position: configPosition, background } = config.theme || {};
-	const { livechatLogo, hideWatermark = false, registrationForm } = config.settings || {};
+	const { color, position: configPosition, background, hideExpandChat } = config.theme || {};
+	const { livechatLogo, hideWatermark = false } = config.settings || {};
 
 	const {
 		color: customColor,
@@ -89,6 +80,7 @@ export const ScreenProvider: FunctionalComponent = ({ children }) => {
 		background: customBackground,
 		hideAgentAvatar = false,
 		hideGuestAvatar = true,
+		hideExpandChat: customHideExpandChat = false,
 	} = iframe.theme || {};
 
 	const [poppedOut, setPopedOut] = useState(false);
@@ -128,7 +120,7 @@ export const ScreenProvider: FunctionalComponent = ({ children }) => {
 	};
 
 	const handleOpenWindow = () => {
-		parentCall('openPopout', store.token);
+		parentCall('openPopout', { token, iframe, customFieldsQueue });
 		dispatch({ undocked: true, minimized: false });
 	};
 
@@ -138,30 +130,14 @@ export const ScreenProvider: FunctionalComponent = ({ children }) => {
 
 	const dismissNotification = () => !isActiveSession();
 
-	const checkPoppedOutWindow = useCallback(async () => {
+	useEffect(() => {
 		// Checking if the window is poppedOut and setting parent minimized if yes for the restore purpose
 		const poppedOut = parse(window.location.search).mode === 'popout';
-		const { token = '' } = parse(window.location.search);
 		setPopedOut(poppedOut);
-
 		if (poppedOut) {
 			dispatch({ minimized: false, undocked: true });
 		}
-
-		if (token && typeof token === 'string') {
-			if (registrationForm && !name && !email) {
-				dispatch({ token });
-				return;
-			}
-			await evaluateChangesAndLoadConfigByFields(async () => {
-				await createOrUpdateGuest({ token });
-			});
-		}
-	}, [dispatch, email, name, registrationForm]);
-
-	useEffect(() => {
-		checkPoppedOutWindow();
-	}, [checkPoppedOutWindow]);
+	}, [dispatch]);
 
 	const screenProps = {
 		theme: {
@@ -174,6 +150,7 @@ export const ScreenProvider: FunctionalComponent = ({ children }) => {
 			background: customBackground || background,
 			hideAgentAvatar,
 			hideGuestAvatar,
+			hideExpandChat: customHideExpandChat || hideExpandChat,
 		},
 		notificationsEnabled: sound?.enabled,
 		minimized: !poppedOut && (minimized || undocked),